+++ confmap/provider/yamlprovider/provider.go
@@ -40,9 +40,9 @@ func New() confmap.Provider {
 	return &provider{}
 }
 
-func (s *provider) Retrieve(_ context.Context, uri string, _ confmap.WatcherFunc) (confmap.Retrieved, error) {
+func (s *provider) Retrieve(_ context.Context, uri string, _ confmap.WatcherFunc) (*confmap.Retrieved, error) {
 	if !strings.HasPrefix(uri, schemeName+":") {
-		return confmap.Retrieved{}, fmt.Errorf("%q uri is not supported by %q provider", uri, schemeName)
+		return nil, fmt.Errorf("%q uri is not supported by %q provider", uri, schemeName)
 	}
 
 	return internal.NewRetrievedFromYAML([]byte(uri[len(schemeName)+1:]))